results/rebuttal_new_baselines/run-20231112-040700_CodeT-SelfRefine-Reflexion-ZeroShot_donnemartin-system-design-oop-finance-donnemartin-system-design-oop-bookclub_0_20-runs_log/finance/1/Reflexion/app.py
from flask import Flask, request
from models import User, Transaction

app = Flask(__name__)

users = {}

@app.route('/create_user', methods=['POST'])
def create_user():
	data = request.get_json()
	user = User(**data)
	users[user.id] = user
	return {'message': 'User created successfully'}, 201

@app.route('/add_transaction/<user_id>', methods=['POST'])
def add_transaction(user_id):
	data = request.get_json()
	transaction = Transaction(**data)
	users[user_id].transactions.append(transaction)
	return {'message': 'Transaction added successfully'}, 201
